+++ packages/mobi-parser/src/mobi.ts
@@ -311,7 +311,9 @@ export class Mobi implements EBookParser {
         URL.revokeObjectURL(url)
       }
       else {
-        unlink(url, () => { })
+        if (existsSync(url)) {
+          unlink(url, () => { })
+        }
       }
     })
     this.resourceCache.clear()
